fix some broken claimReview
The first anomaly of snopes being rated as fake is happening because when I am scraping the Google Fact Check Explorer, for some claimReviews the values of ‘url’ (the fact checking URL in this case snopes.com/…) and ‘itemReviewed.sameAs’ (the URL of the claim) point to the same page. So the solution I am applying is to check when they are the same and remove the ‘claim_url’ from the classified ones. There were 7 occurrences of this anomaly:
- https://www.snopes.com/fact-check/donald-tump-ffa-farmers-president/
https://www.truthorfiction.com/is-united-states-only-country-birthright-citizenship/
- https://www.factcheck.org/2018/10/viral-photo-shows-bill-clinton-with-senator-not-ford/
- https://factcheck.afp.com/no-these-are-not-victims-quake-tsunami-indonesias-sulawesi-island
- https://www.snopes.com/fact-check/pope-francis-married-priests/
- https://www.snopes.com/fact-check/hillary-clinton-hunting-expedition/
- https://www.snopes.com/fact-check/nike-fire-colin-kaepernick-after-arrest/

The Wikipedia anomaly. The cause is that the url of the claim is usually stored in the itemReviewed.author.sameAs property of the ClaimReview schema, but sometimes it just points to the wikipedia page of that Person/Organisation (that should be the correct usage of that property). So the solution I am implementing is to filter out urls belonging to wikipedia in that process. A similar thing happens for some urls belonging to twitter: some sameAs point to twitter.com/<screen_name> while others to twitter.com/<screen_name>/status/<status_id> (fact checking a tweet for example). Solution: filter out the links that point to profiles

BBC/CNN anomaly: the problem is that in the claimReviews they point to the full talk where the fake stuff came out. Example: view-source:https://theconversation.com/what-donald-trump-doesnt-know-about-the-war-of-1812-97997 has a ClaimReview that marks “https://www.cnn.com/2018/06/06/politics/war-of-1812-donald-trump-justin-trudeau-tariff/index.html” as “False” so for these items I don’t really know what to do.